Changes in the build chain, ITK and OpenSSL updates

+++ Horos/Scripts/OpenSSL/Config.sh
@@ -8,10 +8,18 @@ hash="$(find . \( -name CMakeLists.txt -o -name '*.cmake' \) -type f -exec md5 -
 set -e; set -o xtrace
 
 source_dir="$PROJECT_DIR/$TARGET_NAME"
-cmake_dir="$TARGET_TEMP_DIR/CMake"
+cmake_dir="$TARGET_TEMP_DIR/Config"
+install_dir="$TARGET_TEMP_DIR/Install"
 
 mkdir -p "$cmake_dir"; cd "$cmake_dir"
 [ -e "Makefile" -a -f .cmakehash ] && [ "$(cat '.cmakehash')" = "$hash" ] && exit 0
+
+command -v pkg-config >/dev/null 2>&1 || { echo >&2 "error: building $TARGET_NAME requires pkg-config. Please install pkg-config. Aborting."; exit 1; }
+
+mv "$cmake_dir" "$cmake_dir.tmp"
+[ -d "$install_dir" ] && mv "$install_dir" "$install_dir.tmp"
+rm -Rf "$cmake_dir.tmp" "$install_dir.tmp"
+mkdir -p "$cmake_dir"
 
 cd "$cmake_dir"
 rsync -a --delete "$source_dir/" .
@@ -19,8 +27,8 @@ rsync -a --delete "$source_dir/" .
 export CC=clang
 export CXX=clang
 
-config_args=( --prefix="$TARGET_TEMP_DIR/Install" --openssldir="$TARGET_TEMP_DIR/Install" )
-configure_args=( --prefix="$TARGET_TEMP_DIR/Install" --openssldir="$TARGET_TEMP_DIR/Install" )
+config_args=( --prefix="$TARGET_TEMP_DIR/Install" --openssldir="$TARGET_TEMP_DIR/Install" -w )
+configure_args=( --prefix="$TARGET_TEMP_DIR/Install" --openssldir="$TARGET_TEMP_DIR/Install" -w )
 #cfs=($OTHER_CFLAGS)
 #cxxfs=($OTHER_CPLUSPLUSFLAGS)
 
@@ -34,6 +42,9 @@ if [ "$CONFIGURATION" = 'Debug' ]; then
 #    args+=('darwin64-x86_64-cc')
 fi
 
+#cfs+=(-Wno-sometimes-uninitialized)
+#cxxfs+=(-Wno-sometimes-uninitialized)
+
 #if [ ! -z "$CLANG_CXX_LIBRARY" ] && [ "$CLANG_CXX_LIBRARY" != 'compiler-default' ]; then
 #    cxxfs+=(-stdlib="$CLANG_CXX_LIBRARY")
 #fi
@@ -44,11 +55,11 @@ fi
 
 #if [ ${#cfs[@]} -ne 0 ]; then
 #    cfss="${cfs[@]}"
-#    args+=(-DCMAKE_C_FLAGS="$cfss")
+#    configure_args+=(-DCMAKE_C_FLAGS="$cfss")
 #fi
 #if [ ${#cxxfs[@]} -ne 0 ]; then
 #    cxxfss="${cxxfs[@]}"
-#    args+=(-DCMAKE_CXX_FLAGS="$cxxfss")
+#    configure_args+=(-DCMAKE_CXX_FLAGS="$cxxfss")
 #fi
 
 cd "$cmake_dir"
